move utils/api/ to common/api/
