+++ apps/client/src/utils/router.ts
@@ -2,7 +2,7 @@ import { createRouter } from "@tanstack/react-router"
 
 import { routeTree } from "@/routeTree.gen"
 
-import { queryClient } from "./api/query"
+import { queryClient } from "../common/api/query-client"
 
 declare module "@tanstack/react-router" {
     interface Register {
